Move preprocessing of Guild to deserialize_event_with_type
This removes the need to deserialize into a JsonMap, as it can reuse the Value passed there
